Move hasher module to top level
